fix():prod build has assets correctly packaged
