+++ index.html
@@ -15,7 +15,7 @@
   <!-- Global styles -->
   <link href="/global.css" rel="stylesheet" />
 
-  <script type="module" src="./dist/src/manifest-previewer.js"></script>
+  <script type="module" src="/manifest-previewer.js"></script>
 </head>
 
 <body>
