+++ rollup.config.js
@@ -1,32 +1,35 @@
-import resolve from '@rollup/plugin-node-resolve';
-import { terser } from 'rollup-plugin-terser';
-import copy from 'rollup-plugin-copy';
-import html from "@open-wc/rollup-plugin-html";
+import resolve from "@rollup/plugin-node-resolve";
+import { terser } from "rollup-plugin-terser";
+import replace from "@rollup/plugin-replace";
+import strip from "@rollup/plugin-strip";
+import copy from "rollup-plugin-copy";
+import litcss from "rollup-plugin-lit-css";
 
-const config = {
-  input: './index.html',
+export default {
+  input: ["build/manifest-previewer.js"],
   output: {
-    dir: 'build',
-    format: 'es',
-    sourcemap: false
+    file: "dist/manifest-previewer.js",
+    format: "es",
+    sourcemap: false,
   },
   plugins: [
+    resolve(),
+    replace({
+      "process.env.NODE_ENV": JSON.stringify(
+        process.env.NODE_ENV || "production"
+      ),
+    }),
+    litcss(),
+    terser(),
+    strip({
+      functions: ["console.log"],
+    }),
     copy({
       targets: [
-        { src: 'assets/fonts/', dest: 'build/assets/' },
-        { src: 'assets/images/', dest: 'build/assets/' },
-        { src: 'global.css', dest: 'build/' },
-        { src: 'fast-components.min.js', dest: 'build/' },
-        { src: 'dist/src/*.d.ts', dest: 'build/' },
-        { src: 'dist/src/*.d.ts.map', dest: 'build/' }
-      ]
+        { src: "assets/**/*", dest: "dist/assets/" },
+        { src: "global.css", dest: "dist/styles/" },
+        { src: "fast-components.min.js", dest: "dist/" },
+      ],
     }),
-    resolve(),
-    terser(),
-    html()
-  ]
+  ],
 };
-
-export default config;
-
-
